Beam status information was added to the event matching and is now stored in the ANNIEEvent store. The BeamDecoder tool was added to incorporate this information during the EventBuilding process.
    The hit finding process was incorporated into the EventBuilding chain and does not need to be run on the processed files anymore. This reduces the processed file size by a factor of ~100.
    Trigger information about extended readout windows is stored in the ANNIEEvent store now
    It is now possible to build events with 1 VME waveform missing (sometimes the RWM waveform seems to be missing)
    Additional event building options have been introduced, such as Tank+CTC and MRD+CTC (in addition to the standard Tank+MRD+CTC)
    More information is saved in the OrphanStore that can help debug causes for orphaned events.
    The PreProcessTrigOverlap toolchain was added. It should be run before the main EventBuilding process and provides trigger information from the previous and subsequent file that are both needed to match all available MRD+tank timestamps in a single file (the trigger data information overlaps with the previous and subsequent file)
    Timestamp matching cuts have been relaxed from 100ns to 300ns for tank/CTC timestamp matching. Sometimes the timestamps seem to be 256ns apart for no apparent reason (this is the minority of events, but we still want to have them in the data)
    Triggerword matching was extended to triggerword 31 (LED_Start) in addition to triggerword 33 (LED_ADC_Trigger). The LED_Start triggerword timestamp seems to match the tank timestamps much better in beam runs than the LED_ADC_Trigger timestamp.

DataModel changes:

    Two additional members were added to the TriggerClass: The Triggerword and the ExtendedWindow variables
    Two additional labels were added to the MinibufferLabel class, one for MRDCR triggers and one for AmBe events.

Other tool changes:

    The DataSummary tool was extended with more debugging information about orphans and built events
    The LoadANNIEEvent tool was fixed so it can read in normal ANNIEEvent BoostStores again (bugfix). A global event counter option was added in case one reads out multiple subsequent files from the same run and needs a global event numbering scheme.
    The TriggerDataDecoder tool was fixed so it can store multiple triggerwords per timestamp (bugfix)
    The LoadRunInfo tool was added and can be used to load information about the runs from a txt-file (contents for the txt-file can be copied from the SQL database page)
    The PrintANNIEEvent tool was extended to give a better overview of ANNIEEvent data files.

+++ UserTools/ClusterClassifiers/ClusterClassifiers.cpp
@@ -81,7 +81,7 @@ Position ClusterClassifiers::CalculateChargePoint(std::vector<Hit> cluster_hits)
   double tank_center_y = detector_center.Y();
   double tank_center_z = detector_center.Z();
 
-  for (int i = 0; i < cluster_hits.size(); i++){
+  for (int i = 0; i < (int) cluster_hits.size(); i++){
     Hit ahit = cluster_hits.at(i); 
     double hit_charge = ahit.GetCharge();
     int channel_key = ahit.GetTubeId();
@@ -108,7 +108,7 @@ double ClusterClassifiers::CalculateChargeBalance(std::vector<Hit> cluster_hits)
   double total_Q = 0;
   double total_QSquared = 0;
   std::map<int, double> CBMap;
-  for (int i = 0; i < cluster_hits.size(); i++){
+  for (int i = 0; i < (int)cluster_hits.size(); i++){
     Hit ahit = cluster_hits.at(i); 
     double hit_charge = ahit.GetCharge();
     int hit_ID = ahit.GetTubeId();
@@ -133,7 +133,7 @@ double ClusterClassifiers::CalculateChargeBalance(std::vector<Hit> cluster_hits)
 double ClusterClassifiers::CalculateMaxPE(std::vector<Hit> cluster_hits)
 {
   double max_PE = 0;
-  for (int i = 0; i < cluster_hits.size(); i++){
+  for (int i = 0; i < (int)cluster_hits.size(); i++){
     Hit ahit = cluster_hits.at(i); 
     double hit_charge = ahit.GetCharge();
     int channel_key = ahit.GetTubeId();
